refactor project to complify with MVC approach (controller package, model package...)
