+++ TourGuide/src/main/java/com/openclassrooms/tourguide/model/UserPreferences.java
@@ -1,4 +1,4 @@
-package com.openclassrooms.tourguide.user;
+package com.openclassrooms.tourguide.model;
 
 public class UserPreferences {
