+++ TourGuide/src/main/java/com/openclassrooms/tourguide/service/TourGuideService.java
@@ -3,8 +3,8 @@ package com.openclassrooms.tourguide.service;
 import com.openclassrooms.tourguide.dto.NearByAttraction;
 import com.openclassrooms.tourguide.helper.InternalTestHelper;
 import com.openclassrooms.tourguide.tracker.Tracker;
-import com.openclassrooms.tourguide.user.User;
-import com.openclassrooms.tourguide.user.UserReward;
+import com.openclassrooms.tourguide.model.User;
+import com.openclassrooms.tourguide.model.UserReward;
 
 import java.time.LocalDateTime;
 import java.time.ZoneOffset;
